+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # Changelog
+* (2016-06-12). Option not to display the run data on the post was added. Option to display the run data on the excerpt was added.
 * (2016-06-11). Initial Release.